pyproject.toml

[tool.poetry]
name = "tcp_extractor"
version = "1.0.0"
description = "TCP Extractor"
authors = ["CogniteJohan <johan.dahl@cognite.com>"]

[tool.black]
line-length = 120
target_version = ['py37']
include = '\.py$'

[tool.isort]
line_length=120                # corresponds to -w  flag
multi_line_output=3            # corresponds to -m  flag
include_trailing_comma=true    # corresponds to -tc flag
skip_glob = '^((?!py$).)*$'    # this makes sort all Python files
known_third_party = ["cognite", "dotenv", "psutil"]

[tool.poetry.dependencies]
python = "^3.7"
cognite-sdk = "^2.32.0"
cognite-extractor-utils = "^1.5.4"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pyinstaller = "^4.0"
macholib = {version = "^1.14", platform = "darwin"}             # Used by pyinstaller pn Mac OS
pywin32-ctypes = {version = "^0.2.0", platform = "win32"}       # Used by pyinstaller on Windows
pefile = "^2019.4.18"                                           # Used by pyinstaller on Windows
mypy = "^0.910"
flake8 = "^4.0.1"
black = "^21.11b1"
isort = "^5.10.1"
pre-commit = "^2.15.0"

[tool.poetry.scripts]
tcp_extractor = "tcp_extractor.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
